feat: Update admin routes to include authentication prefix and add product validation schema
