+++ src/pages/Admin/products/pages/new/page.jsx
@@ -22,8 +22,7 @@ import FeaturesSection from "./sections/features";
 import DimensionsWeightSection from "./sections/dimensions-weight";
 import TagsKeywordsSection from "./sections/tags-keywords";
 import AdditionalInformationSection from "./sections/additional-information";
-
-import * as z from "zod";
+import { productSchema } from "@/utils/product_zod";
 
 const defaultCategories = [];
 
@@ -31,76 +30,6 @@ const defaultCategories = [];
 const CLOUDINARY_UPLOAD_PRESET = "products"; // Replace with your upload preset
 const CLOUDINARY_CLOUD_NAME = "drv6qpv56";
 const CLOUDINARY_UPLOAD_URL = `https://api.cloudinary.com/v1_1/${CLOUDINARY_CLOUD_NAME}/image/upload`;
-
-const productSchema = z.object({
-  id: z.string().min(8, "ID must be at least 8 characters long"),
-  title: z.string().min(6, "Title must be at least 6 characters long"),
-  price: z.number().min(1, "Price must be greater than 0"),
-  originalPrice: z.number().min(1, "Original price must be greater than 0"),
-  thumbnail: z.string().min(2, "Thumbnail must be at least 2 characters long"),
-  description: z
-    .string()
-    .min(10, "Description must be at least 10 characters long"),
-  category: z.string().min(2, "Category must be at least 2 characters long"),
-  stock: z.number().min(0, "Stock must be a non-negative number"),
-  brand: z.string().min(2, "Brand must be at least 2 characters long"),
-  tax: z.number().default(0),
-  tags: z
-    .array(z.string())
-    .min(3, "At least three tag is required")
-    .default([]),
-  keywords: z.array(z.string()).optional().default([]),
-  detailedDescription: z
-    .string()
-    .min(10, "Detailed description must be at least 10 characters long")
-    .max(1200, "Detailed description must be at most 1200 characters long")
-    .default(""),
-  specifications: z
-    .array(
-      z.object({
-        label: z.string().min(2, "Label must be at least 2 characters long"),
-        value: z.string().min(2, "Value must be at least 2 characters long"),
-      }),
-    )
-    .default([]),
-  features: z
-    .array(z.string())
-    .min(3, "At least three feature is required")
-    .default([]),
-  images: z
-    .array(z.string())
-    .min(1, "At least one image is required")
-    .default([]),
-  rating: z.number().optional().default(0),
-  reviewCount: z.number().optional().default(0),
-  dimensions: z
-    .object({
-      width: z.number().optional().default(0),
-      height: z.number().optional().default(0),
-      depth: z.number().optional().default(0),
-    })
-    .optional(),
-  weight: z.number().optional().default(0),
-  createdAt: z.date().default(Date.now),
-  updatedAt: z.date().default(Date.now),
-  isActive: z.boolean().default(true),
-  currency: z.string().optional().default("INR"),
-  owner: z
-    .object({
-      id: z.string().default("admin"),
-      name: z.string().default("Admin"),
-    })
-    .optional(),
-  warranty: z.string().optional().default(null),
-  returnPolicy: z.string().optional().default(null),
-  shippingDetails: z.string().optional().default(null),
-  relatedProducts: z.array(z.string()).optional().default([]),
-  accessories: z.array(z.string()).optional().default([]),
-  priceHistory: z.array(z.number()).optional().default([]),
-  sku: z.string().min(2, "SKU must be at least 2 characters long").default(""),
-  views: z.number().default(0),
-  specialInfo: z.record(z.any()).optional().default({}),
-});
 
 export default function NewProductPage() {
   const navigate = useNavigate();
@@ -445,7 +374,7 @@ export default function NewProductPage() {
 
       <div className="flex items-center gap-4">
         <Button variant="ghost" size="icon" asChild>
-          <Link to="/admin/products">
+          <Link to="/auth/admin/products">
             <ArrowLeft className="h-4 w-4" />
           </Link>
         </Button>
@@ -579,7 +508,7 @@ export default function NewProductPage() {
             {/* Submit Button */}
             <div className="flex justify-end gap-4">
               <Button type="button" variant="outline" asChild>
-                <Link to="/admin/products">Cancel</Link>
+                <Link to="/auth/admin/products">Cancel</Link>
               </Button>
               <Button type="submit" disabled={isSubmitting}>
                 {isSubmitting ? (
